# serverside/swagger_server/models/courses.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Courses(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, name: str=None, description: str=None, semester: int=None, year: int=None, credits: float=None, insti_id: float=None, slot_id: float=None):  # noqa: E501
        """Courses - a model defined in Swagger

        :param id: The id of this Courses.  # noqa: E501
        :type id: int
        :param name: The name of this Courses.  # noqa: E501
        :type name: str
        :param description: The description of this Courses.  # noqa: E501
        :type description: str
        :param semester: The semester of this Courses.  # noqa: E501
        :type semester: int
        :param year: The year of this Courses.  # noqa: E501
        :type year: int
        :param credits: The credits of this Courses.  # noqa: E501
        :type credits: float
        :param insti_id: The insti_id of this Courses.  # noqa: E501
        :type insti_id: float
        :param slot_id: The slot_id of this Courses.  # noqa: E501
        :type slot_id: float
        """
        self.swagger_types = {
            'id': int,
            'name': str,
            'description': str,
            'semester': int,
            'year': int,
            'credits': float,
            'insti_id': float,
            'slot_id': float
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'description': 'description',
            'semester': 'semester',
            'year': 'year',
            'credits': 'credits',
            'insti_id': 'insti_id',
            'slot_id': 'slot_id'
        }
        self._id = id
        self._name = name
        self._description = description
        self._semester = semester
        self._year = year
        self._credits = credits
        self._insti_id = insti_id
        self._slot_id = slot_id

    @classmethod
    def from_dict(cls, dikt) -> 'Courses':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Courses of this Courses.  # noqa: E501
        :rtype: Courses
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Courses.


        :return: The id of this Courses.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Courses.


        :param id: The id of this Courses.
        :type id: int
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Courses.


        :return: The name of this Courses.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Courses.


        :param name: The name of this Courses.
        :type name: str
        """

        self._name = name

    @property
    def description(self) -> str:
        """Gets the description of this Courses.


        :return: The description of this Courses.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this Courses.


        :param description: The description of this Courses.
        :type description: str
        """

        self._description = description

    @property
    def semester(self) -> int:
        """Gets the semester of this Courses.


        :return: The semester of this Courses.
        :rtype: int
        """
        return self._semester

    @semester.setter
    def semester(self, semester: int):
        """Sets the semester of this Courses.


        :param semester: The semester of this Courses.
        :type semester: int
        """

        self._semester = semester

    @property
    def year(self) -> int:
        """Gets the year of this Courses.


        :return: The year of this Courses.
        :rtype: int
        """
        return self._year

    @year.setter
    def year(self, year: int):
        """Sets the year of this Courses.


        :param year: The year of this Courses.
        :type year: int
        """

        self._year = year

    @property
    def credits(self) -> float:
        """Gets the credits of this Courses.


        :return: The credits of this Courses.
        :rtype: float
        """
        return self._credits

    @credits.setter
    def credits(self, credits: float):
        """Sets the credits of this Courses.


        :param credits: The credits of this Courses.
        :type credits: float
        """

        self._credits = credits

    @property
    def insti_id(self) -> float:
        """Gets the insti_id of this Courses.


        :return: The insti_id of this Courses.
        :rtype: float
        """
        return self._insti_id

    @insti_id.setter
    def insti_id(self, insti_id: float):
        """Sets the insti_id of this Courses.


        :param insti_id: The insti_id of this Courses.
        :type insti_id: float
        """

        self._insti_id = insti_id

    @property
    def slot_id(self) -> float:
        """Gets the slot_id of this Courses.


        :return: The slot_id of this Courses.
        :rtype: float
        """
        return self._slot_id

    @slot_id.setter
    def slot_id(self, slot_id: float):
        """Sets the slot_id of this Courses.


        :param slot_id: The slot_id of this Courses.
        :type slot_id: float
        """

        self._slot_id = slot_id