# serverside/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Learnfinity API
  description: API Documentation for Learnfinity
  version: 1.0.0
servers:
- url: /
...
paths:
  /signup:
    post:
      tags:
      - auth
      summary: Signup a new user
      operationId: signup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User-signup'
        required: true
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /signin:
    post:
      tags:
      - auth
      summary: Signin an existing user
      operationId: signin_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User-signin'
        required: true
      responses:
        "201":
          description: User logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /forgotpwd:
    post:
      tags:
      - auth
      summary: Reset password
      operationId: forgotpwd_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User-email'
        required: true
      responses:
        "201":
          description: Password reset
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /logout:
    post:
      tags:
      - auth
      summary: User logged out
      operationId: logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User-email'
        required: true
      responses:
        default:
          description: User logged out
      x-openapi-router-controller: swagger_server.controllers.auth_controller
...
  /courses:
    get:
      tags:
      - courses
      summary: Get all courses that the user is enrolled in
      operationId: courses_get
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Courses'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.courses_controller
    post:
      tags:
      - courses
      summary: Create a new course *(admin)
      operationId: courses_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Courses-create'
        required: true
      responses:
        "201":
          description: Newly created course object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Courses'
      x-openapi-router-controller: swagger_server.controllers.courses_controller
...
  /courses/{year}:
    get:
      tags:
      - courses
      summary: Get all courses in that year
      operationId: courses_year_get
      parameters:
      - name: year
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Courses'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.courses_controller
  /courses/{year}/{semester}:
    get:
      tags:
      - courses
      summary: Get all courses in that year
      operationId: courses_year_semester_get
      parameters:
      - name: year
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      - name: semester
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Courses'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.courses_controller
  /courses/{year}/{semester}/{slot_id}:
    get:
      tags:
      - courses
      summary: Get all courses in that year
      operationId: courses_year_semester_slot_id_get
      parameters:
      - name: year
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      - name: semester
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      - name: slot_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Courses'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.courses_controller
  /courses/edit:
    post:
      tags:
      - courses
      summary: Edit course information *(admin)
      operationId: courses_edit_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Courses-edit'
        required: true
      responses:
        default:
          description: Course object with edited information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Courses'
      x-openapi-router-controller: swagger_server.controllers.courses_controller
  /courses/adduser:
    post:
      tags:
      - courses
      summary: Add a user to the course *(admin & faculty)
      operationId: courses_adduser_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Courses-user'
        required: true
      responses:
        default:
          description: User added succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course-User'
      x-openapi-router-controller: swagger_server.controllers.courses_controller
  /courses/remuser:
    post:
      tags:
      - courses
      summary: Remove a user from the course *(admin & faculty)
      operationId: courses_remuser_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Courses-user'
        required: true
      responses:
        default:
          description: User removed succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course-User'
      x-openapi-router-controller: swagger_server.controllers.courses_controller
  /evaluation:
    get:
      tags:
      - evaluation
      summary: Get all the evaluations from the courses user is enrolled in
      operationId: evaluation_get
      responses:
        "200":
          description: List of evaluations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evaluation'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.evaluation_controller
    post:
      tags:
      - evaluation
      summary: Create an evaluation *(admin & faculty)
      operationId: evaluation_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evaluation-create'
        required: true
      responses:
        "201":
          description: Evaluation object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluation'
      x-openapi-router-controller: swagger_server.controllers.evaluation_controller
  /evaluation/{course_id}:
    get:
      tags:
      - evaluation
      summary: Get all the evaluations from the course specified
      operationId: evaluation_course_id_get
      parameters:
      - name: course_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: List of evaluations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evaluation'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.evaluation_controller
  /evaluation/edit:
    post:
      tags:
      - evaluation
      summary: Edit an evaluation
      operationId: evaluation_edit_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evaluation'
        required: true
      responses:
        default:
          description: Evaluation object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluation'
      x-openapi-router-controller: swagger_server.controllers.evaluation_controller
  /evaluation/delete:
    post:
      tags:
      - evaluation
      summary: Delete an evaluation
      operationId: evaluation_delete_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evaluation-id'
        required: true
      responses:
        default:
          description: Evaluation deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluation'
      x-openapi-router-controller: swagger_server.controllers.evaluation_controller
...
components:
  schemas:
    User:
      type: object
      properties:
        email_id:
          type: string
        insti_id:
          type: integer
        is_Admin:
          type: boolean
        is_Prof:
          type: boolean
        profile_pic:
          type: string
      example:
        email_id: email_id
        is_Prof: true
        profile_pic: profile_pic
        is_Admin: true
        insti_id: 0
    User-signup:
      type: object
      properties:
        email_id:
          type: string
        password:
          type: string
        insti_id:
          type: integer
        is_Prof:
          type: boolean
        profile_pic:
          type: string
    User-signin:
      type: object
      properties:
        email_id:
          type: string
        password:
          type: string
    User-email:
      type: object
      properties:
        email_id:
          type: string
...
    Courses:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        semester:
          type: integer
        year:
          type: integer
        credits:
          type: number
        insti_id:
          type: number
        slot_id:
          type: number
      example:
        year: 1
        credits: 5.962133916683182377482808078639209270477294921875
        slot_id: 2.3021358869347654518833223846741020679473876953125
        name: name
        description: description
        semester: 6
        id: 0
        insti_id: 5.63737665663332876420099637471139430999755859375
    Courses-create:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        semester:
          type: integer
        year:
          type: integer
        credits:
          type: number
        slot_id:
          type: number
    Courses-edit:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        semester:
          type: integer
        year:
          type: integer
        credits:
          type: number
        slot_id:
          type: number
    Courses-user:
      type: object
      properties:
        course_id:
          type: number
        user_id:
          type: number
    Course-User:
      type: object
      properties:
        id:
          type: integer
        course_id:
          type: number
        user_id:
          type: number
      example:
        course_id: 6.02745618307040320615897144307382404804229736328125
        user_id: 1.46581298050294517310021547018550336360931396484375
        id: 0
    Evaluation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        staticfile_id:
          type: string
        deadline:
          type: string
        course_id:
          type: number
        weightage:
          type: number
        total_marks:
          type: number
        submission_allowed:
          type: boolean
      example:
        weightage: 1.46581298050294517310021547018550336360931396484375
        course_id: 6.02745618307040320615897144307382404804229736328125
        total_marks: 5.962133916683182377482808078639209270477294921875
        staticfile_id: staticfile_id
        id: 0
        title: title
        deadline: deadline
        submission_allowed: true
    Evaluation-create:
      type: object
      properties:
        title:
          type: string
        staticfile_id:
          type: string
        deadline:
          type: string
        course_id:
          type: number
        weightage:
          type: number
        total_marks:
          type: number
        submission_allowed:
          type: boolean
    Evaluation-id:
      type: object
      properties:
        id:
          type: integer
...